Revert "Nyoba ngirim frame"
This reverts commit 866c9f6577ea0cde2d9bd31f0c60236e673ca124.

+++ slidingwindow.h
@@ -37,7 +37,7 @@ typedef struct FRAME {
  	unsigned int stx;
  	Byte *data;
  	unsigned int etx;
- 	short checksum;
+ 	Byte checksum;
 } FRAME;
 
 typedef struct ACKN {
